refactor: LazyDynamicSequence を splay 木実装に一本化
実測 (judge.yosupo.jp/submission/384939 vs 384962) で、AVL 版
(join-based framework) は reverse/apply のような split して戻す操作の
再結合コストが splay 版より重く、Time (全ケース中の最大値) で
1.783s -> 1.073s (約40%減) の差が出た。同一問題に対する重複実装を
避けるため、LazyDynamicSequence の中身を splay_dynamic_sequence.hpp の
実装（親ポインタを持つ反復版 splay）に置き換え、後者は削除。
公開 API・型名は変更なし。

verify (test/yosupo/data_structure/dynamic_sequence_range_affine_range_sum
.test.cpp) は既存のまま LazyDynamicSequence を使用しているため変更不要。
重複していた *_splay.test.cpp は削除。

ローカルで自前fuzzテスト・max_right/min_leftテスト・Library Checker
サンプル・small.cppとの小規模突き合わせ・wrong_avl_killer/
wrong_splay_killer(4シード)/extreme_insertion/max_random の全出力が
splay版と一致することを確認済み。
